refactor: header 내의 title과 closeIcon을 서브 컴포넌트로 분리
