+++ components/src/lib/modalCloseIcon/ModalCloseIcon.module.css
@@ -1,0 +1,6 @@
+.closeButton {
+    width: 24px;
+    height: 24px;
+    object-fit: contain;
+    cursor: pointer;
+}
