+++ components/src/lib/modalHeader/ModalHeader.tsx
@@ -1,37 +1,11 @@
 import styles from './ModalHeader.module.css';
-import CloseIcon from '../assets/closeButton.svg';
 
-interface ModalHeaderProps extends React.HTMLAttributes<HTMLElement> {
-  onClose: () => void;
-  title?: string;
-  showCloseIcon?: boolean;
-  customCloseIcon?: string;
-}
+interface ModalHeaderProps extends React.HTMLAttributes<HTMLElement> {}
 
-const ModalHeader = ({
-  onClose,
-  title,
-  showCloseIcon = true,
-  customCloseIcon,
-  style,
-  ...rest
-}: React.PropsWithChildren<ModalHeaderProps>) => {
-  const onErrorIcon = (event: React.SyntheticEvent<HTMLImageElement, Event>) => {
-    event.currentTarget.src = CloseIcon;
-  };
-
+const ModalHeader = ({ style, children, ...rest }: React.PropsWithChildren<ModalHeaderProps>) => {
   return (
     <header className={styles.modalHeader} style={style} {...rest}>
-      <span className={styles.title}>{title && title}</span>
-      {showCloseIcon && (
-        <img
-          src={customCloseIcon ?? CloseIcon}
-          alt="close"
-          className={styles.closeButton}
-          onClick={onClose}
-          onError={onErrorIcon}
-        />
-      )}
+      {children}
     </header>
   );
 };
