+++ components/src/lib/modalTitle/ModalTitle.module.css
@@ -1,0 +1,8 @@
+.title {
+    color: black;
+    font-family: Noto Sans KR;
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 26.06px;
+    text-align: left;
+}
